// frontend/src/api-generated/models/dotKiemKeResponse.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * IT Asset Management API
 * Tài liệu API Hệ thống quản lý tài sản CNTT
 * OpenAPI spec version: 1.0
 */

export interface DotKiemKeResponse {
  id?: number;
  idDonVi?: number;
  maDotKiemKe?: string;
  tenDotKiemKe?: string;
  tenNguoiLap?: string;
  tenNguoiPheDuyet?: string;
  thoiGianBatDauDuKien?: string;
  thoiGianKetThucDuKien?: string;
  thoiGianThucHien?: string;
  thoiGianChotSoLieu?: string;
  trangThai?: string;
  tongTaiSanHeThong?: number;
  tongTaiSanThucTe?: number;
  thoiGianTao?: string;
  thoiGianCapNhat?: string;
}